pyproject.toml
[tool.poetry]
name = "prefiracode"
version = "0.1.0"
description = "Pretendard + Fira Code"
authors = ["Joonmo Yang <dev@remagpie.com>"]
license = "OFL-1.1"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.2"
fonttools = "^4.38.0"
pillow = "^9.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
